+++ test/fixtures/cookbooks/test/recipes/add_to_list.rb
@@ -3,16 +3,45 @@ cookbook_file '/tmp/dangerfile3' do
   mode '00644'
 end
 
-# add first entry. 1 delim, ends_with
-add_to_list 'Add to list 1' do
+# test lists with an item seperator
+add_to_list 'Add to an empty list, seperator' do
   path '/tmp/dangerfile3'
   pattern 'empty_list='
   delim [' ']
   entry 'newentry'
 end
 
-# # add an entry to a existing list. 1 delim, ends_with
-add_to_list 'Add to DEFAULT_APPEND' do
+add_to_list 'Add an entry to an empty list, seperator' do
+  path '/tmp/dangerfile3'
+  pattern 'last_list='
+  delim [' ']
+  entry 'single'
+end
+
+add_to_list 'Add a duplicate entry to a list, seperator' do
+  path '/tmp/dangerfile3'
+  pattern 'People to call:'
+  delim [', ']
+  entry 'Bobby'
+end
+
+add_to_list 'Add a new entry to a list, seperator' do
+  path '/tmp/dangerfile3'
+  pattern 'People to call:'
+  delim [', ']
+  entry 'Harry'
+end
+
+# test lists with an item seperator and terminal list string
+add_to_list 'Add to an empty list, seperator, terminal' do
+  path '/tmp/dangerfile3'
+  pattern 'DEFAULT_APPEND_EMPTY='
+  delim [' ']
+  ends_with '"'
+  entry 'first'
+end
+
+add_to_list 'Add an existing item a list, seperator, terminal' do
   path '/tmp/dangerfile3'
   pattern 'DEFAULT_APPEND='
   delim [' ']
@@ -20,8 +49,7 @@ add_to_list 'Add to DEFAULT_APPEND' do
   entry 'showopts'
 end
 
-# add an entry to a existing list. 1 delim, ends_with
-add_to_list 'Add to DEFAULT_APPEND again' do
+add_to_list 'Add a new item to a list, seperator, terminal' do
   path '/tmp/dangerfile3'
   pattern 'DEFAULT_APPEND='
   delim [' ']
@@ -29,8 +57,30 @@ add_to_list 'Add to DEFAULT_APPEND again' do
   entry 'addtogrub'
 end
 
-# add first entry. 2 delim, ends_with
-add_to_list 'Add to list using Regexp escaped input' do
+# test lists with an item seperator, item delimiters
+add_to_list 'Add an item to an empty list, seperator and item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo2d_empty='
+  delim [',', '"']
+  entry 'single'
+end
+
+add_to_list 'Add an existing entry to a list, seperator and item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo2d_list'
+  delim [',', '"']
+  entry 'first2'
+end
+
+add_to_list 'Add a new entry to a list, seperator and item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo2d_list'
+  delim [',', '"']
+  entry 'third2'
+end
+
+# test lists with an item seperator, item delimiters and a terminal list string
+add_to_list 'Add to an empty list with an item seperator, item delimiters and a terminal list string' do
   path '/tmp/dangerfile3'
   pattern Regexp.escape('empty_delimited_list=(')
   delim [', ', '"']
@@ -38,128 +88,65 @@ add_to_list 'Add to list using Regexp escaped input' do
   entry 'newentry'
 end
 
-# # add existing entry. 2 delim, ends_with
-# add_to_list 'Operation 19' do
-#   path '/tmp/dangerfile3'
-#   pattern Regexp.escape('last_delimited_list= (')
-#   delim [',', '|']
-#   ends_with ')'
-#   entry 'single'
-# end
-#
-# # Add existing entry list. 2 delim, ends_with
-# add_to_list 'Operation 20' do
-#   path '/tmp/dangerfile3'
-#   pattern 'my @net1918 ='
-#   delim [', ', '"']
-#   ends_with ');'
-#   entry '172.16.0.0/12'
-# end
-#
-# # Add new entry to a non empty list. 2 delim, ends_with
-# add_to_list 'Operation 21' do
-#   path '/tmp/dangerfile3'
-#   pattern 'my @net1918 ='
-#   delim [', ', '"']
-#   ends_with ');'
-#   entry '33.33.33.0/24'
-# end
-#
-# # add first entry. 3 delim, ends_with
-# add_to_list 'Operation 22' do
-#   path '/tmp/dangerfile3'
-#   pattern Regexp.escape('empty_3delim=(')
-#   delim [', ', '[', ']']
-#   ends_with ')'
-#   entry 'newentry'
-# end
-#
-# # add existing entry. 3 delim, ends_with
-# add_to_list 'Operation 23' do
-#   path '/tmp/dangerfile3'
-#   pattern 'multi '
-#   delim [', ', '[', ']']
-#   ends_with ')'
-#   entry '818'
-# end
-#
-# # add new entry to non-empty list. 3 delim, ends_with
-# add_to_list 'Operation 24' do
-#   path '/tmp/dangerfile3'
-#   pattern 'multi '
-#   delim [', ', '[', ']']
-#   ends_with ')'
-#   entry '323'
-# end
-#
-# # add first entry. 1 delim
-# add_to_list 'Operation 25' do
-#   path '/tmp/dangerfile3'
-#   pattern 'last_list='
-#   delim [' ']
-#   entry 'single'
-# end
-#
-# # add existing entry. 1 delim
-# add_to_list 'Operation 26' do
-#   path '/tmp/dangerfile3'
-#   pattern 'People to call:'
-#   delim [', ']
-#   entry 'Bobby'
-# end
-#
-# # add a new entry. 1 delim
-# add_to_list 'Operation 27' do
-#   path '/tmp/dangerfile3'
-#   pattern 'People to call:'
-#   delim [', ']
-#   entry 'Harry'
-# end
-#
-# # add first entry. 2 delim
-# add_to_list 'Operation 28' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo2d_empty='
-#   delim [',', '"']
-#   entry 'single'
-# end
-#
-# # add existing entry. 2 delim
-# add_to_list 'Operation 29' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo2d_list'
-#   delim [',', '"']
-#   entry 'first2'
-# end
-#
-# # add a new entry. 2 delim
-# add_to_list 'Operation 30' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo2d_list'
-#   delim [',', '"']
-#   entry 'third2'
-# end
-#
-# # add first entry. 3 delim
-# add_to_list 'Operation 31' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo3d_empty='
-#   delim [',', '[', ']']
-#   entry 'single'
-# end
-#
-# # add existing entry. 3 delim
-# add_to_list 'Operation 32' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo3d_list'
-#   delim [',', '[', ']']
-#   entry 'first3'
-# end
-#
-# # add a new entry. 3 delim
-# add_to_list 'Operation 33' do
-#   path '/tmp/dangerfile3'
-#   pattern 'wo3d_list'
-#   delim [',', '[', ']']
-#   entry 'third3'
-# end
+add_to_list 'Add an existing entry to a list with an item seperator, item delimiters and a terminal list string' do
+  path '/tmp/dangerfile3'
+  pattern Regexp.escape('last_delimited_list= (')
+  delim [',', '|']
+  ends_with ')'
+  entry 'single'
+end
+
+add_to_list 'Add an existing entry to a list with item seperator, item delimiters and a terminal list string' do
+  path '/tmp/dangerfile3'
+  pattern 'my @net1918 ='
+  delim [', ', '"']
+  ends_with ');'
+  entry '172.16.0.0/12'
+end
+
+add_to_list 'Add an entry to a list with item seperator, item delimiters and a terminal list string' do
+  path '/tmp/dangerfile3'
+  pattern 'my @net1918 ='
+  delim [', ', '"']
+  ends_with ');'
+  entry '33.33.33.0/24'
+end
+
+add_to_list 'Add an existing item to a complex list' do
+  path '/tmp/dangerfile3'
+  pattern 'multi '
+  delim [', ', '[', ']']
+  ends_with ')'
+  entry '818'
+end
+
+# test lists with an item seperator, before and after item delimiters
+add_to_list 'Add first entry to a list, seperator, item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo3d_empty='
+  delim [',', '[', ']']
+  entry 'single'
+end
+
+add_to_list 'Add an existing entry, seperator, item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo3d_list'
+  delim [',', '[', ']']
+  entry 'first3'
+end
+
+add_to_list 'Add a new entry, seperator, item delimiters' do
+  path '/tmp/dangerfile3'
+  pattern 'wo3d_list'
+  delim [',', '[', ']']
+  entry 'third3'
+end
+
+# test lists with an item seperator, before and after item delimiters and a terminal list string
+add_to_list 'Add to list using Regexp escaped input' do
+  path '/tmp/dangerfile3'
+  pattern Regexp.escape('empty_3delim=(')
+  delim [' ', '[', ']']
+  ends_with ')'
+  entry 'newentry'
+end