+++ test/fixtures/cookbooks/test/recipes/delete_lines_missing_file.rb
@@ -1,0 +1,4 @@
+delete_lines 'missing_file' do
+  path '/tmp/nofilehere'
+  pattern '^#.*'
+end